requirements.txt
PyExifTool==0.5.5
